Fixes type error in makeSVG
N needs to be explicitly cast to int
in order to make np.random.choice happy.

+++ deadleaves/python/makeSVG.py
@@ -116,7 +116,7 @@ def scaleEmojis(emojis,alpha=0.01,xmin=0.01,xLim=100,yLim=100):
 def getN(alpha):
 	"""assumes each element has diameter equal to scale, which isn't really true,
 	but the approximation actually works fairly well"""
-	return (alpha**-1)*(10**3.6987)
+	return int((alpha**-1)*(10**3.6987))
 
 def getAlpha(intensity):
 	"""in case we decide to use a different scale for intensity"""
